This commit introduces support for tab and newline characters in the … (#113)
* This commit introduces support for tab and newline characters in the DrawingContextImpl class. The update will handle these characters during draw operations, translating tab characters into a sequence of spaces and newline characters into a 'no-op' (no operation) at this stage.

* Refactor drawing context implementation

The unused namespace `System.Text` was removed from the `DrawingContextImpl` class. Also, the handling of tab spaces was adjusted; instead of incrementing the current position one by one, we now increment it by the `tabSize` in one go. This also involved creating a constant for the `tabSize`. Additionally, comments related to resharper's handling of closure were added.

* Automated JetBrains cleanup

* Implement clipping in DrawingContextImpl

The implementation of clipping was added in the DrawingContextImpl.cs. This ensures the pixel buffer coordinates stay within the defined bounds when rendering text, specifically with tabs, thereby improving the rendering process and accuracy.