Features * Added Italic support * Turned on Strikethrough (although it renders as underline) * Added Canvas gallery page * Removed dependency on Crayon library
Bug Fixes
* Fix line draws 1 too many chars
* Fix underline draws 1 too many underline

Changes:
* Remove dependency on Crayon library, as it didn't implement italics and added ConsoleUtils
* Use Italic escape to handle FontStyle=Italic
* Refactor DrawLineInternal into  DrawLineInternal and DrawRectangleLineInternal

Details:
The issue is that DrawLineInternal was written to draw rectangle borders and then simply reused as the DrawLine implementation.
DrawLineRectangle draws Corner + Line + Corner. This always ended up drawing 1 more than the desired line length, which is appropriate for borders and rectangles.

The solution was to refactor the code into
* DrawLineInternal
* DrawRectangleLineInternal