Fix strikethrough implementation (#134)
* Fix strikethrough implementation

Details:
I didn't have a way of determing when strikethrough vs underline was happening.

I finally figured a somewhat hacky but legit as far as I can tell way to signal underline and strikethrough, which is to use a large value of thickness in the font metric.

Since we only support a thickness of 1 in a console app, this feels like a safe way to do it and shouldn't get in the way of anything.

* add default to switch with exception

* fix lint issue

* Automated JetBrains cleanup

Co-authored-by:  <+@users.noreply.github.com>

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>