Fix for #438 (#439)
#438 fix

Pen width as already accounted for in the rect, removed the addition of the pen width.
Correctly renders rectangle now
<img width="543" height="524" alt="image" src="https://github.com/user-attachments/assets/31d5f704-f049-40af-9eb2-1267e97cf4e9" />

Also renamed GalleryAnimatedLines => GalleryAnimation as it animates mulitple things now.

Also cleaned up access patterns by adding to PixelBuffer[Point] which uses already existing PixelBufferCoordinate cast. This significantly cleans up the access patter in DrawContextImpl, as we can now pass Point (which we need for clip comparison) directly to the buffer without the explicit cast.  The code is much cleaner to read.

Based on AI feedback I also got rid of allocation of array of SKColors in DrawBitmap